// models/violationsDetails/violationsDetails.ts
export interface ViolationsDetails {
    id: number
    driver_id: number
    violation_report_number: string
    violation_state: string
    violation_count: number
    violation_category: string
    inspection_level: number
    out_of_service: string
    total_points: number
    document_id: number
    total_violation_severity: number
    violation_severity: number
    violation_group: string
    carrier_id: number
    driver_name: string
    time_weight: number
    violation_date: string
    violation_sequence: number
    oos_violation_severity: number
    violation_description: string
    updated_date: string
    vehicle_id: number
  }